+++ packages/plugin-inquirer-cli/test/Core.mocks.js
@@ -49,6 +49,8 @@ class CoreMock {
         },
         collections: {
           selected: "foo-current",
+          ids: [],
+          plain: [],
         },
       },
       server: {